Use run and i64 instead of eval and INT for examples.

+++ examples/custom_types_and_methods.rs
@@ -1,8 +1,8 @@
-use rhai::{Engine, EvalAltResult, INT};
+use rhai::{Engine, EvalAltResult};
 
 #[derive(Debug, Clone)]
 struct TestStruct {
-    x: INT,
+    x: i64,
 }
 
 impl TestStruct {
